Add route, route-test generator.
Note that we're intentionally skipping a couple edge-case-ish scenarios
here, because they're sufficiently rare, and this resolution issue seems
to be kind of buggy on the Ember CLI side, that it's not worth pursuing
further at present.

Also, we do not (and cannot) generate or update the `router.ts` file at
present; we leave it as JavaScript. We'll need upstream changes to
ember-cli to support that.