feature(Octane): remove all blueprint files
The new Octane-ready blueprints will work by transforming the output of
the default generators: by renaming the file from `.js` to `.ts` and by
transforming the AST for files where necessary (e.g. adding a reference
to a registry). The net is that we will no longer use standalone files
for these, which in turn means we will have much less work to do to stay
in sync with the upstream.

(It also means that if we come to a point where the upstream so desires,
it will be able to run tests to see if changes there break this repo!)