Add test for helper-addon blueprint back.
The helpers blueprint behaves rather differently than the other
blueprints, it seems. First, it still has traces of MU. If I rename the
`__collection__` folder to `__path__`, the `--pod` tests fail. If I
remove the explicit `helper-addon` blueprint, the helper-addon test and
addon-related helper tests fail, despite the fact that the corresponding
tests for `initializer-addon` and `route-addon` *do* work.

Once the `helper-addon` blueprint can be removed, then we will also be
able to remove `--addon-import.js`.